Created the index and home page. Added html designs and created some txt files with color schemes and other useful information. Setup anime, and watch list on home page, created the login and signup buttons on the page. The Login Page and Sign up page arent being used because of modals.
Next is confirmation page and a forgotpassword page for signup and login, the watchlist page and more brainstorming. I connected all the href's on the home and index page to their respective pages in the directory.

Beware home may need the references to have ../ since what it's referencing may be in a parent folder.